feat(board-service): 댓글 생성시 포인트 적립을 위해 comment.created 토픽에 이벤트 발행
- 게시글 수정, 삭제 API 구현
- 댓글 수정, 삭제 API 구현
- 게시글, 댓글은 생성일 기준 하루가 지나야 수정 & 삭제 가능

+++ board-service/src/main/java/woobl0g/boardservice/comment/service/CommentService.java
@@ -65,11 +65,28 @@ public class CommentService {
         if(!comment.getUser().getUserId().equals(userId)) {
             throw new CommentException(ResponseCode.COMMENT_DELETE_FORBIDDEN);
         }
+        if (!comment.canModify()) {
+            throw new CommentException(ResponseCode.COMMENT_MODIFY_TOO_EARLY);
+        }
 
         if (!comment.getChildren().isEmpty()) {
             comment.softDelete();
         } else {
             commentRepository.delete(comment);
         }
+    }
+
+    @Transactional
+    public void update(Long commentId, CreateCommentRequestDto dto, Long userId) {
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new CommentException(ResponseCode.COMMENT_NOT_FOUND));
+        if(!comment.getUser().getUserId().equals(userId)) {
+            throw new CommentException(ResponseCode.COMMENT_UPDATE_FORBIDDEN);
+        }
+        if(!comment.canModify()) {
+            throw new CommentException(ResponseCode.COMMENT_MODIFY_TOO_EARLY);
+        }
+
+        comment.update(dto.getContent());
     }
 }
